basic modification graph class
When running through the curation protocol, I noticed that it would be useful to have a history of all of the changes made to a particular structure. Every molecule will have one of these graphs to track its history. It will be linear in most cases, but can branch when two structures are merged, and their histories will be combined (likely only during when deduplicating). I'd like to be able to actually view it as a graph in the eventual GUI, but for right now it's just a text string describing each change.

I might include smiles/inchi for each node in the graph, so each node can have a drawing.

I still need to actually include the modification graph in the Mol class.